API partial refactoring
