+++ api/views/__init__.py
@@ -1,0 +1,4 @@
+from .crawler import CrawlerViewSet
+from .crawler_instance import CrawlerInstanceViewSet
+from .crawler_queue import CrawlerQueueViewSet
+from .task import TaskViewSet
