+++ crawler_manager/settings.py
@@ -3,7 +3,7 @@ import os
 # Kafka host information
 KAFKA_TOPIC_PREFIX = os.getenv('KAFKA_TOPIC_PREFIX', 'crawler_ufmg')
 
-KAFKA_HOSTS = [x.strip() for x in os.getenv('KAFKA_HOSTS', 'kafka:9092').split(',')]
+KAFKA_HOSTS = [x.strip() for x in os.getenv('KAFKA_HOSTS', 'localhost:9092').split(',')]
 KAFKA_CONSUMER_AUTO_OFFSET_RESET = 'earliest'
 KAFKA_CONSUMER_TIMEOUT = 120000
 KAFKA_CONSUMER_COMMIT_INTERVAL_MS = 5000
@@ -16,7 +16,7 @@ KAFKA_REQUEST_TIMEOUT_MS = 5 * 60 * 1000
 KAFKA_SESSION_TIMEOUT_MS = 2 * 60 * 1000
 
 # Redis host information
-REDIS_HOST = os.getenv('REDIS_HOST', 'redis')
+REDIS_HOST = os.getenv('REDIS_HOST', 'localhost')
 REDIS_PORT = int(os.getenv('REDIS_PORT', 6379))
 REDIS_DB = int(os.getenv('REDIS_DB', 0))
 REDIS_PASSWORD = os.getenv('REDIS_PASSWORD', None)
@@ -39,7 +39,7 @@ LINK_GENERATOR_TOPIC = os.getenv('LINK_GENERATOR_TOPIC', KAFKA_TOPIC_PREFIX + '.
 WRITER_TOPIC = os.getenv('WRITER_TOPIC', KAFKA_TOPIC_PREFIX + '.writer')
 
 STOPPED_SPIDER_NOTIFICATION_ADDRESS = os.getenv(
-    'STOPPED_SPIDER_NOTIFICATION_ADDRESS', 'http://web:8000/detail/stop_crawl/{crawler_id}')
+    'STOPPED_SPIDER_NOTIFICATION_ADDRESS', 'http://localhost:8000/detail/stop_crawl/{crawler_id}')
 
 TASK_TOPIC = os.getenv('TASK_TOPIC', KAFKA_TOPIC_PREFIX + 'task_topic')
 TASK_DATA_CONSUMER_GROUP = os.getenv('TASK_DATA_CONSUMER_DATA', KAFKA_TOPIC_PREFIX + '.task_data_group')
