+++ interface/settings.py
@@ -44,7 +44,7 @@ if SECRET_KEY is None:
     SECRET_KEY = get_random_secret_key()
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = env('DEBUG')
+DEBUG = True#env('DEBUG')
 
 ALLOWED_HOSTS = env('DJANGO_ALLOWED_HOSTS')
 
@@ -61,6 +61,7 @@ INSTALLED_APPS = [
     'crispy_forms',
     'rest_framework',
     'mathfilters',
+    'api.apps.ApiConfig'
 ]
 
 CRISPY_TEMPLATE_PACK = 'bootstrap4'
