+++ main/staticfiles/json/steps_signature.json
@@ -1,1 +1,1 @@
-[{"name": "break_image_captcha", "mandatory_params": ["xpath_input", "xpath_output"], "optional_params": {"preprocessing": null}}, {"name": "clique", "mandatory_params": ["xpath"], "optional_params": {}}, {"name": "cssify", "mandatory_params": ["xpath"], "optional_params": {}}, {"name": "digite", "mandatory_params": ["xpath", "texto"], "optional_params": {}}, {"name": "element_in_page", "mandatory_params": ["xpath"], "optional_params": {}}, {"name": "espere", "mandatory_params": ["segs"], "optional_params": {}}, {"name": "for_clicavel", "mandatory_params": ["xpath"], "optional_params": {}}, {"name": "gera_nome_arquivo", "mandatory_params": [], "optional_params": {}}, {"name": "nesse_elemento_esta_escrito", "mandatory_params": ["xpath", "texto"], "optional_params": {}}, {"name": "opcoes", "mandatory_params": ["xpath"], "optional_params": {"exceto": null}}, {"name": "pegue_os_links_da_paginacao", "mandatory_params": ["xpath_dos_botoes", "xpath_dos_links"], "optional_params": {"indice_do_botao_proximo": -1}}, {"name": "print_", "mandatory_params": ["word"], "optional_params": {}}, {"name": "range_", "mandatory_params": ["stop"], "optional_params": {}}, {"name": "salva_pagina", "mandatory_params": [], "optional_params": {}}, {"name": "selecione", "mandatory_params": ["xpath", "opcao"], "optional_params": {}}, {"name": "wait_page", "mandatory_params": [], "optional_params": {}}]
+[{"name": "clique", "name_display": "Clicar", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["elemento"], "optional_params": {}, "field_options": {}}, {"name": "comparacao", "name_display": "Compara\u00e7\u00e3o", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["arg1", "comp", "arg2"], "optional_params": {}, "field_options": {"comp": {"field_type": "select", "select_options": ["'=='", "'<='", "'>='", "'<'", "'>'", "'!='"]}}}, {"name": "digite", "name_display": "Digitar em", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["xpath", "texto"], "optional_params": {}, "field_options": {}}, {"name": "elemento_existe_na_pagina", "name_display": "Checar se elemento existe na p\u00e1gina", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["xpath"], "optional_params": {}, "field_options": {}}, {"name": "espere", "name_display": "Esperar", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["segundos"], "optional_params": {}, "field_options": {"segundos": {"field_type": "number", "input_placeholder": "Espera em segundos"}}}, {"name": "extrai_texto", "name_display": "Extrair texto de", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["xpath"], "optional_params": {}, "field_options": {}}, {"name": "for_clicavel", "name_display": "\u00c9 clic\u00e1vel", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["xpath"], "optional_params": {}, "field_options": {}}, {"name": "imprime", "name_display": "Imprimir", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["texto"], "optional_params": {}, "field_options": {}}, {"name": "localiza_elementos", "name_display": "Localizar elementos", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["xpath"], "optional_params": {"numero_xpaths": null, "modo": "Modo Simples"}, "field_options": {"modo": {"field_type": "select", "select_options": ["'Modo Simples'", "'XPath Complexos'"]}}}, {"name": "nesse_elemento_esta_escrito", "name_display": "Est\u00e1 escrito", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["xpath", "texto"], "optional_params": {}, "field_options": {}}, {"name": "objeto", "name_display": "Objeto", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["objeto"], "optional_params": {}, "field_options": {}}, {"name": "opcoes", "name_display": "Op\u00e7\u00f5es", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["xpath"], "optional_params": {"exceto": null}, "field_options": {}}, {"name": "quebrar_captcha_imagem", "name_display": "Quebrar captcha de imagem", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["xpath_do_elemento_captcha", "xpath_do_campo_a_preencher"], "optional_params": {"funcao_preprocessamento": null}, "field_options": {}}, {"name": "repete", "name_display": "Repetir", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["vezes"], "optional_params": {}, "field_options": {}}, {"name": "retorna_pagina", "name_display": "Voltar", "executable_contexts": ["page", "tab"], "mandatory_params": [], "optional_params": {}, "field_options": {}}, {"name": "run_javascript", "name_display": "Executar c\u00f3digo Javascript", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["codigo_javascript"], "optional_params": {}, "field_options": {}}, {"name": "salva_pagina", "name_display": "Salvar p\u00e1gina", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": [], "optional_params": {}, "field_options": {}}, {"name": "selecione", "name_display": "Selecionar", "executable_contexts": ["page", "tab", "iframe"], "mandatory_params": ["xpath", "opcao"], "optional_params": {}, "field_options": {}}]
